requirements.txt
absl-py==0.7.0
astor==0.7.1
backports.weakref==1.0.post1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Deprecated==1.2.4
enum34==1.1.6
et-xmlfile==1.0.1
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
funcsigs==1.0.2
futures==3.2.0
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
lightgbm==2.2.2
Markdown==3.0.1
MarkupSafe==1.1.0
mock==2.0.0
nltk==3.4
numpy==1.16.0
openpyxl==2.5.14
pandas==0.24.0
pbr==5.1.1
pickle-mixin==1.0.2
protobuf==3.6.1
PyGithub==1.43.5
PyJWT==1.7.1
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
SQLAlchemy==1.2.17
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
textblob==0.15.2
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
xgboost==0.81
xlrd==1.2.0
